Moves `update_collection_metadata` transient entity to `home/collections/update` folder
