+++ lib/application/pages/home/collections/update/update_collection_details.dart
@@ -7,6 +7,7 @@ import 'package:memo/application/constants/exception_strings.dart';
 import 'package:memo/application/constants/strings.dart' as strings;
 import 'package:memo/application/hooks/rich_text_field_controller_hook.dart';
 import 'package:memo/application/hooks/tags_controller_hook.dart';
+import 'package:memo/application/pages/home/collections/update/update_collection_metadata.dart';
 import 'package:memo/application/theme/theme_controller.dart';
 import 'package:memo/application/view-models/home/update_collection_details_vm.dart';
 import 'package:memo/application/view-models/home/update_collection_vm.dart';
@@ -15,7 +16,6 @@ import 'package:memo/application/widgets/theme/rich_text_field.dart';
 import 'package:memo/application/widgets/theme/tags_field.dart';
 import 'package:memo/application/widgets/unfocus_pointer.dart';
 import 'package:memo/core/faults/exceptions/base_exception.dart';
-import 'package:memo/domain/transients/update_collection_metadata.dart';
 import 'package:memo/domain/validators/collection_validators.dart' as validators;
 
 class UpdateCollectionDetails extends ConsumerWidget {
