+++ lib/application/pages/home/collections/update/update_collection_memos.dart
@@ -7,13 +7,13 @@ import 'package:memo/application/constants/animations.dart' as anims;
 import 'package:memo/application/constants/dimensions.dart' as dimens;
 import 'package:memo/application/constants/images.dart' as images;
 import 'package:memo/application/constants/strings.dart' as strings;
+import 'package:memo/application/pages/home/collections/update/update_collection_metadata.dart';
 import 'package:memo/application/pages/home/collections/update/update_memo_terminal.dart';
 import 'package:memo/application/theme/theme_controller.dart';
 import 'package:memo/application/view-models/home/update_collection_memos_vm.dart';
 import 'package:memo/application/view-models/home/update_collection_vm.dart';
 import 'package:memo/application/widgets/material/asset_icon_button.dart';
 import 'package:memo/application/widgets/theme/rich_text_field.dart';
-import 'package:memo/domain/transients/update_collection_metadata.dart';
 
 class UpdateCollectionMemos extends HookConsumerWidget {
   @override
