+++ lib/application/view-models/home/update_collection_memos_vm.dart
@@ -1,8 +1,8 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:memo/application/pages/home/collections/update/update_collection_metadata.dart';
 import 'package:memo/application/pages/home/collections/update/update_collection_providers.dart';
-import 'package:memo/domain/transients/update_collection_metadata.dart';
 
 final updateCollectionMemosVM = StateNotifierProvider.autoDispose<UpdateCollectionMemosVM, UpdateMemosState>(
   (ref) => UpdateCollectionMemosVMImpl(memos: ref.read(updateMemosMetadata)),
