+++ lib/application/view-models/home/update_collection_vm.dart
@@ -2,9 +2,9 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:memo/application/pages/home/collections/update/update_collection_metadata.dart';
 import 'package:memo/application/pages/home/collections/update/update_collection_providers.dart';
 import 'package:memo/core/faults/exceptions/base_exception.dart';
-import 'package:memo/domain/transients/update_collection_metadata.dart';
 import 'package:memo/domain/validators/collection_validators.dart';
 
 final updateCollectionVM = StateNotifierProvider.autoDispose<UpdateCollectionVM, UpdateCollectionState>(
